pyproject.toml
[project]
name = "job-recommender"
version = "0.1.0"
description = "GitHub profile-based job recommendation system using Perplexity AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "google-cloud-aiplatform>=1.38.0",
    "perplexityai>=0.1.0",
    "PyGithub>=2.1.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]

